Update tutorial-create-instance.md
Cloud sync documentation clearly states that it does not support AADDS when synchronizing users in a comparison table between AAD Connect and AAD Cloud sync. Had a case for a customer who did not know this point and it is worth to mention in this document. https://docs.microsoft.com/en-us/azure/active-directory/cloud-sync/what-is-cloud-sync#comparison-between-azure-ad-connect-and-cloud-sync